requirements.txt
fastapi
uvicorn
python-telegram-bot[ext]
python-dotenv